backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
openai-whisper==20231117
sounddevice==0.4.7
numpy==1.26.4
pyttsx3==2.98
gTTS==2.5.4
playsound==1.3.0
langchain-anthropic
langchain-core
